Update to use base game's validators (instead of using old copy without updated features, e.g. sharing a stack of food (since like 1.1))
Now postfix SpawnedFoodSearchInnerScan, which is called inside BestFoodSourceOnMap instead of a prefix that bypassed it all.

Adding setting for the max distance to use room food.
Postfix uses found food to increase max distance for table to be as far as the found food.

No longer supporting obsolete SmarterFoodSelection.cs